Replace requests with aiohttp and make warpy asynchronous
This is a major change in how this overlay will be defined as it will now be asynchronous. After doing
a bit of research this is likely to be the best option since there will be various functions polling
the API for information. If one requests takes to long the GUI will get hung up, this resolves that
issue.

Note: This is a first dive into asynchronous so future commits may aim to resolve sloppy asynchronous
implementations.