Clean up timer widget code and create state widgets more intelligently
This commit corresponds to the most recent update of warpy and removes the need
for state handlers in order to determine the string value of the open worlds
state. This also further abstracts the code for the asynchronous state_widget
function by taking the function and timer widget instead of specifying these
explicitly in code. This removes the need to have three near identical timer
state setting functions due to a small difference in function calls despite
ingesting the same target information, just with different strings and values.